requirements.txt
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
decorator==5.2.1
deep-translator==1.11.4
filelock==3.18.0
fsspec==2025.7.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20250625
pillow==11.3.0
proglog==0.1.12
python-dotenv==1.1.1
regex==2025.7.34
requests==2.32.4
soupsieve==2.7
sympy==1.14.0
tiktoken==0.9.0
ffmpeg-python==0.2.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
